requirements.txt
streamlit
lightgbm
numpy
pandas
pillow
scikit-learn